Restructure into package; add tests.
- We removed some output images that were committed by accident.
- Added documentation to `tracing.split_picture`
- Added a doctest for `split_picture`, and fixed bug uncovered in the
  process.
- Moved `pipeline.py` to `pipeline_click.py`; this is an old pipeline
  and should probably be removed.
- Replaced some TABs by spaces.

With Théo, Sara, and Dennis.
