+++ butterfly/binarization.py
