+++ butterfly/measurement.py
