+++ butterfly/ruler_detection.py
